feat(#2687): add lazy iter_utxos + iter_identities_with_metadata, address #2692 review
Rebased onto current development after #2690 landed. Combines #2692's
original contribution (the new iterator API surface) with the review
fixes Copilot flagged on the original PR.

## New API surface (iter_utxos + iter_identities_with_metadata)

`BlockchainStore::iter_utxos`
  - Lazy iterator: `Iterator<Item = StorageResult<(OutPoint, Utxo)>>`.
  - Wraps the sled tree iter directly so rows yield one at a time —
    hydrate paths can stream into projection tables without holding
    the full UTXO set in memory (mainnet-scale: unbounded).
  - Default impl errors; SledStore provides the real scan.

`BlockchainStore::iter_identities_with_metadata`
  - Companion to dev's existing `iter_identities` (consensus-only,
    eager). This variant pairs each consensus row with its optional
    `IdentityMetadata` (display name, did string, controlled nodes)
    and is the surface the startup hydrate path uses to rebuild
    `identity_registry` from sled.
  - Lazy + `Result` items for the same reason as iter_utxos.
  - Renamed from `iter_identities` to avoid a name clash with dev's
    pre-existing version (#2639 migration).

## #2692 review fixes (5/5 from Copilot)

L29 — `HotStateProjectionMeta` count fields: `usize` → `u64`.
  bincode is serialized to disk; `usize` round-trips differently on
  32-bit vs 64-bit targets and can fail deserialize on the smaller
  arch. Fixed-width avoids the portability hazard.

L103 — `stage_employment_contracts` looped the entire registry on
  every `CreateEmploymentContract` tx → O(historical contract count)
  per block. Replaced with `stage_employment_contracts_for_block`
  which filters by `start_height == height`, so only contracts created
  in the current block get re-staged. Older contracts are already in
  the projection from when they originally committed.

L143 — `stage_contract_blocks` same pattern: O(total) per block,
  unbounded batch growth. Replaced (at the per-block staging site)
  with `stage_contract_blocks_for_block` which filters by
  `block_height == height`. The original `stage_contract_blocks`
  helper is kept for the one-time replay-time backfill rebuild.

L327 + L1470 — atomicity gap. `commit_metadata_write` applies per-tree
  batches without cross-tree WAL atomicity (its doc-comment is
  explicit: "This path is not WAL-protected"). A crash mid-commit
  could leave `projection_hot_state_meta` durable while sibling
  projection tables are stale → `hot_state_projection_is_current()`
  returns true on restart and hydrates an inconsistent hot state.

  Fix: split meta into a separate commit that runs AFTER the data
  batch flushes. New helper
    `commit_hot_state_projection_meta_only(store, block)`
  opens its own `begin_metadata_write` → stage → `commit_metadata_write`
  cycle. Meta only becomes durable once the data tables are durable.

  Worst case after crash: meta absent → fast path skipped on next
  startup → fallback to historical replay → backfill re-stages
  atomically. Never the meta-current + data-stale combination.

  Wired in:
    * `stage_hot_state_projection_updates` no longer stages meta.
    * Executor block-apply path (blockchain.rs) calls
      `commit_hot_state_projection_meta_only` AFTER
      `commit_metadata_write` succeeds.
    * Pure-legacy non-executor block-apply path keeps meta+data
      bundled because `commit_block` IS WAL-protected — explicit
      stage call added at the same site.
    * Mixed executor/legacy fn: stages meta only when
      `!using_executor` to avoid double-writing meta on the
      executor path (which writes meta via the separate batch).
    * `backfill_hot_state_projections_from_replay`: data committed
      first, then meta committed separately.

## Tests

  cargo test -p lib-blockchain --lib test_iter_
    -> 7/7 pass:
       test_iter_utxos_returns_committed_unspent_outputs,
       test_iter_utxos_empty_store_yields_nothing,
       test_iter_utxos_returns_all_entries,
       test_iter_identities_with_metadata_returns_consensus_and_metadata,
       test_iter_identities_with_metadata_handles_missing_metadata,
       test_iter_identities_with_metadata_empty_store_yields_nothing,
       test_iter_observer_records_returns_all_admitted (existing)

  cargo test -p lib-blockchain --lib test_hot_state_projection
    -> 1/1 pass:
       test_hot_state_projection_tables_commit_atomically (existing)

+++ lib-blockchain/src/projections.rs
@@ -18,14 +18,18 @@ pub struct HotStateProjectionMeta {
     pub height: u64,
     pub block_hash: [u8; 32],
     pub completed_at_unix: u64,
-    pub validators: usize,
-    pub gateways: usize,
-    pub domains: usize,
-    pub credentials: usize,
-    pub employment_contracts: usize,
-    pub dao_entries: usize,
-    pub pouw_mints: usize,
-    pub contract_blocks: usize,
+    // Reviewer #2692/L29: fixed-width counts so the bincode payload is
+    // architecture-independent. `usize` round-trips differently between
+    // 32-bit and 64-bit targets and can fail deserialization if a count
+    // ever exceeds the local `usize` width.
+    pub validators: u64,
+    pub gateways: u64,
+    pub domains: u64,
+    pub credentials: u64,
+    pub employment_contracts: u64,
+    pub dao_entries: u64,
+    pub pouw_mints: u64,
+    pub contract_blocks: u64,
 }
 
 impl HotStateProjectionMeta {
